refactor(chords): split chordUtils into focused modules following SRP
Extract three single-responsibility modules from the monolithic chordUtils:
- chord-transposition.ts: NOTES constant, transposeNote helper, transposeChord
- key-conversion.ts: semitonesToKeyName, songKeyToSemitones
- chord-detection.ts: isChordLine, parseChordFile

chordUtils.ts is kept as a barrel re-export so all existing imports
continue to work without changes.

Test files are split accordingly:
- chord-transposition.test.ts (16 tests, includes slash-chord music-theory suite)
- key-conversion.test.ts (11 tests)
- chord-detection.test.ts (4 tests)
